refactor(benchmarks): isolate presentation workload examples
